+++ server/.env
@@ -1,4 +1,4 @@
-SECRET_KEY=b'\xa9\x10\xe7ba\xc0eM\xae\x1b\xd31_\xa3<U'
+SECRET_KEY='8e1399f88795c973babbc72e4f3ea3b4b7a15e7269c11de891d843eb8d759d68'
 
 
 # SQL Database
@@ -10,12 +10,7 @@ SQL_USER="the_hive"
 SQL_PASSWORD="Phds%40Knust"
 SQL_DATABASE="phds"
 
-
 # sqlite Database
-SQL_LITE_PATH="phds.db"
+SQL_LITE_PATH="pdhs.db"
+INSTANCE_PATH="C:\Users\Ayarmz\Documents\Projects\Papperless-Document-Handling-System/server/instance"
 
-
-# a = 'dialect+driver://username:password@host:port/database'
-# 'dialect+driver://username:password@host:port/database'
-# mssql+pyodbc://the_hive%40the-hive:Phds%40Knust@the-hive.database.windows.net/phds?driver=SQL+Server
-# mssql+pymssql://the_hive:Phds%40Knust@the-hive.database.windows.net/phds
