+++ server/config/development.py
@@ -4,12 +4,15 @@ from os import environ, path
 from dotenv import load_dotenv
 
 
-basedir = path.abspath(path.dirname(__file__))
-load_dotenv(path.join(basedir, '.env'))
+load_dotenv()
+
 
 SECRET_KEY = environ.get('SECRET_KEY')
 
 FLASK_ENV = 'development'
 DEBUG = True
 SESSION_COOKIE_SECURE = False
-DATABASE_URI = f"sqlite:///{environ.get('SQL_LITE_PATH')}"
+DATABASE_URI = "sqlite:///{instance_path}/{path}".format(
+    instance_path=environ.get('INSTANCE_PATH'),
+    path=environ.get('SQL_LITE_PATH')
+)
