+++ server/database.py
@@ -1,45 +1,58 @@
 import sqlite3
 from flask_sqlalchemy import SQLAlchemy
+from sqlalchemy import MetaData
 import click
 from flask import current_app, g
 from flask.cli import with_appcontext
+from flask_migrate import Migrate
 
-db = SQLAlchemy()
-
-
-def get_db():
-    if 'db' not in g:
-        g.db = sqlite3.connect(
-            current_app.config['DATABASE'],
-            detect_types=sqlite3.PARSE_DECLTYPES
-        )
-        g.db.row_factory = sqlite3.Row
-
-    return g.db
+metadata = MetaData(
+    naming_convention={
+        'pk': 'pk_%(table_name)s',
+        'fk': 'fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s',
+        'ix': 'ix_%(table_name)s_%(column_0_name)s',
+        'uq': 'uq_%(table_name)s_%(column_0_name)s',
+        'ck': 'ck_%(table_name)s_%(constraint_name)s',
+    }
+)
 
 
-def close_db(e=None):
-    db = g.pop('db', None)
+db = SQLAlchemy(metadata=metadata)
+migrate = Migrate()
 
-    if db is not None:
-        db.close()
+# def get_db():
+#     if 'db' not in g:
+#         g.db = sqlite3.connect(
+#             current_app.config['DATABASE'],
+#             detect_types=sqlite3.PARSE_DECLTYPES
+#         )
+#         g.db.row_factory = sqlite3.Row
 
-
-def init_db():
-    db = get_db()
-
-    with current_app.open_resource('schema.sql') as f:
-        db.executescript(f.read().decode('utf8'))
-
-
-@click.command('init-db')
-@with_appcontext
-def init_db_command():
-    """Clear the existing data and create new tables."""
-    init_db()
-    click.echo('Initialized the database.')
+#     return g.db
 
 
-def init_app(app):
-    app.teardown_appcontext(close_db)
-    app.cli.add_command(init_db_command)
+# def close_db(e=None):
+#     db = g.pop('db', None)
+
+#     if db is not None:
+#         db.close()
+
+
+# def init_db():
+#     db = get_db()
+
+#     with current_app.open_resource('schema.sql') as f:
+#         db.executescript(f.read().decode('utf8'))
+
+
+# @click.command('init-db')
+# @with_appcontext
+# def init_db_command():
+#     """Clear the existing data and create new tables."""
+#     init_db()
+#     click.echo('Initialized the database.')
+
+
+# def init_app(app):
+#     app.teardown_appcontext(close_db)
+#     app.cli.add_command(init_db_command)
