+++ server/pdhs_app/templates/home.html
@@ -1,19 +1,19 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>MySQL Testing</title>
-</head>
-<body>
+  </head>
+  <body>
     <form method="POST">
-        <p>Name <input type="text" name="username"></p>
-        <p>Password <input type="password" name="password"></p>
-        <p><input type="submit" value="Submit"></p>
+      <p>Name <input type="text" name="username" /></p>
+      <p>Password <input type="password" name="password" /></p>
+      <p><input type="submit" value="Submit" /></p>
     </form>
     {% if user %}
-    <h3> Welcome {{user.password}}</h3>
+    <h3>Welcome {{user.password}}</h3>
     {% endif %}
-</body>
+  </body>
 </html>
