+++ server/pdhs_app/templates/test.html
@@ -1,29 +1,27 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Test Document</title>
-</head>
-<body>
-    <h3> Something Here </h3>
+  </head>
+  <body>
+    <h3>Something Here</h3>
     {% if users %}
     <table>
-        <tr>
-        {% if keys %}
-            {% for key in keys %}
-                <th> {{key}} </th>
-            {% endfor %}
-        {% endif %}
-        </tr>
-        {% for user in users %}
-        <tr>
-            <td> {{user.name}} </td> 
-            <td> {{user.password}} </td> 
-        </tr>
-        {% endfor %}
+      <tr>
+        {% if keys %} {% for key in keys %}
+        <th>{{key}}</th>
+        {% endfor %} {% endif %}
+      </tr>
+      {% for user in users %}
+      <tr>
+        <td>{{user.username}}</td>
+        <td>{{user.email}}</td>
+      </tr>
+      {% endfor %}
     </table>
     {% endif %}
-</body>
+  </body>
 </html>
