+++ server/run.py
@@ -1,6 +1,6 @@
 from pdhs_app import create_app
 
-app = create_app(env='testing')
+app = create_app(env='development')
 
 if __name__ == '__main__':
-    app.run()
+    app.run(host='0.0.0.0')
